move ets related code into a separate lib
