+++ web/models/dadi.ex
@@ -9,6 +9,7 @@ defmodule Efl.Dadi do
   alias Efl.Dadi
   alias Efl.TimeUtil
   alias Efl.Mailer
+  alias Efl.EtsManager
   require IEx
 
   schema "dadi" do
@@ -21,45 +22,42 @@ defmodule Efl.Dadi do
     timestamps()
   end
 
-  @ets_table :dadi
-  @ets_key :is_running
+  #def start(ref_category) do
+    #try do
+      #IO.puts("Deleting all records")
+      #Repo.delete_all(Dadi)
 
-  def start(ref_category) do
-    try do
-      IO.puts("Deleting all records")
-      Repo.delete_all(Dadi)
+      #IO.puts("Start fetching categories")
+      #ref_category
+      #|> Category.create_items
 
-      IO.puts("Start fetching categories")
-      ref_category
-      |> Category.create_items
+      #IO.puts("Start fetching posts")
+      #Post.update_contents 
 
-      IO.puts("Start fetching posts")
-      Post.update_contents 
+      #IO.puts("Exporting Xls file")
+      #Efl.Xls.Dadi.create_xls
 
-      IO.puts("Exporting Xls file")
-      Efl.Xls.Dadi.create_xls
-
-      IO.puts("Sending Emails")
-      Mailer.send_email_with_xls 
-    rescue
-      _ -> Mailer.send_alert
-    end
-  end
+      #IO.puts("Sending Emails")
+      #Mailer.send_email_with_xls 
+    #rescue
+      #_ -> Mailer.send_alert
+    #end
+  #end
 
   def start do
     try do
-      ets_create_table
-      if ets_lookup do
+      EtsManager.ets_create_table
+      if EtsManager.ets_lookup do
         IO.puts("The app is running")
       else
-        ets_insert(true)
+        EtsManager.ets_insert(true)
         main()
-        ets_insert(false)
+        EtsManager.ets_insert(false)
       end
     rescue
       ex ->
         IO.inspect(ex)
-        Mailer.send_alert(ex)
+        #Mailer.send_alert(ex)
     end
   end
 
@@ -105,32 +103,6 @@ defmodule Efl.Dadi do
     else
       changeset
     end 
-  end
-
-  #Todo move ets functions to an independent file
-  defp ets_create_table do
-    try do
-      :ets.new(@ets_table, [:set, :protected, :named_table])
-    rescue
-      _ -> IO.puts("The ets table exists!")
-    end
-  end
-
-  def ets_insert(is_running) do
-    try do
-      :ets.insert(@ets_table, { @ets_key, is_running })
-    rescue
-      _ -> IO.puts("Ets insert fail")
-    end
-  end
-
-  defp ets_lookup do
-    try do
-      { @ets_key, value } = :ets.lookup(@ets_table, @ets_key) |> List.first
-      value
-    rescue
-      _ -> false
-    end
   end
 
   defp main do
